services/api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
kafka-python==2.0.2
psycopg2-binary==2.9.9